requirements/bionemo_requirements.txt
nvidia-bionemo-sdk
requests
python-dotenv
